new table builder program
This program is made to build either the flat table for the CFD people and the interpolation table. This commit has the listbox gui setup to select which case to run. It also has the beginnings of the data read in done for the flat table. The structure of the interpolation table needs to be clarified, been over a month since that discussion took place.