interpolation_app_V0.0
First shippable version of the interpolation app completed. App interpolates forces for body and wheel of the MQS in shallow water conditions. Heading, Depth, and Speed are on sliders with real time changing spinner to see what the value is and to dial in exact values. All inputs have limits so unexpected values are rejected. User uses a radio dial to select the wheel steering angle, in future version of the app this will use a quadratic interpolation to give the user access to values between those tested. Rough test shows everything works, need to do a QA test to see if it can be made to function erroneously. Also if you enter exact values in the table, because it is using scattered interpolation it will not give you the exact value back. This can be fixed in future deployments, right now see an error on order of 1-2 N in output for body forces.