interpolation_app V0.2
Now has the ability to linearly interpolate on the wheel angle. Going to try a quadratic interpolation, but this proves the mechanics/structure of the method works.

Added a constant to the ARC_table_builder to deal with the rounding. Right now any value higher than 2 breaks the function. It will still finish successfully but the table will be unusable and error out almost every time. This is the big item to fix going forward.