arcInterpolator function
Added ARC_Interpolation_Table.mat to the github so that the app and interpolator function may be used. Mild update to app eliminated one line of redundant code.

arcInterpolator.m created. This is a standalone function that can be used in other scripts to get interpolated forces and moments from a function created by the experimental sample points. Function is currently untested but reuses much of the working code from the app. Next update will be tested and shippable code.